fix(schema): fix nullable union with types
