+++ packages/specs/schema/src/domain/JsonSchema.ts
@@ -652,12 +652,13 @@ export class JsonSchema extends Map<string, any> {
           this.set("minLength", undefined);
           break;
         case null:
-          this.any(
-            ...["null"].concat(
-              this.getJsonType(),
-              this.$allow.map((v) => typeof v)
-            )
-          );
+          this.nullable(true);
+          // this.any(
+          //   ...["null"].concat(
+          //     this.getJsonType(),
+          //     this.$allow.map((v) => typeof v)
+          //   )
+          // );
           break;
       }
     });
@@ -1270,9 +1271,6 @@ export class JsonSchema extends Map<string, any> {
 
   protected setManyOf(keyword: "oneOf" | "anyOf" | "allOf", value: (AnyJsonSchema | null)[]) {
     let resolved = value
-      // .map((o) => {
-      //   return o === null ? {type: "null"} : o;
-      // })
       .filter((o) => {
         if ((o && "type" in o && o.type === "null") || o === null) {
           this.nullable(true);
